refactor(proxmox): change VMID type from int to uint64 across Proxmox provider
Updates VMID parameter and field types from int to uint64 throughout the Proxmox provider implementation,
including API request structures, provider structs, client methods, and LXC-related functions.
Also updates string conversion calls from strconv.Itoa to strconv.FormatUint.

+++ internal/api/v1/proxmox/common.go
@@ -2,5 +2,5 @@ package proxmoxapi
 
 type ActionRequest struct {
 	Node string `uri:"node" binding:"required"`
-	VMID int    `uri:"vmid" binding:"required"`
+	VMID uint64 `uri:"vmid" binding:"required"`
 } //	@name	ProxmoxVMActionRequest
